photos-server/pyproject.toml
[tool.poetry]
name = "Ruurd Photos"
version = "2.0.0"
description = ""
authors = ["Ruurd Bijlsma <ruurd@bijlsma.dev>"]
license = "'"
readme = "README.md"
packages = [{ include = "*", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = { extras = ["all"], version = "^0.115.0" }
sqlalchemy = { extras = ["mypy"], version = "^2.0.35" }
uvicorn = "^0.31.0"
httpx = "^0.27.2"
authlib = "^1.3.2"
itsdangerous = "^2.2.0"
pillow = "^10.4.0"
psycopg2-binary = "^2.9.9"
watchdog = "^5.0.3"
tqdm = "^4.66.5"
pre-commit = "^3.8.0"
types-tqdm = "^4.66.0.20240417"
pydantic-settings = "^2.5.2"
alembic = "^1.13.3"
pyyaml = "^6.0.2"
types-pyyaml = "^6.0.12.20240917"
pydantic = {extras = ["dotenv"], version = "^2.9.2"}
passlib = "^1.7.4"
python-jose = "^3.3.0"
oauthlib = "^3.2.2"
piexif = "^1.1.3"


[tool.poetry.group.dev.dependencies]
mypy = "^1.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"